requirements.txt
apache-beam[gcp]==2.37.0
google-cloud-storage==2.1.0
google-cloud-firestore==2.4.0
